Move DebuggerHandler to a separate file
